+++ modules/repository/src/test/scala/com/fortysevendeg/repository/app/AppRepositorySpec.scala
@@ -135,8 +135,8 @@ trait AppMockCursor
     (category, 4, appSeq map (_.data.category), StringDataType),
     (imagePath, 5, appSeq map (_.data.imagePath), StringDataType),
     (colorPrimary, 6, appSeq map (_.data.colorPrimary), StringDataType),
-    (dateInstalled, 7, appSeq map (_.data.dateInstalled), DoubleDataType),
-    (dateUpdate, 8, appSeq map (_.data.dateUpdate), DoubleDataType),
+    (dateInstalled, 7, appSeq map (_.data.dateInstalled), LongDataType),
+    (dateUpdate, 8, appSeq map (_.data.dateUpdate), LongDataType),
     (version, 9, appSeq map (_.data.version), StringDataType),
     (installedFromGooglePlay, 10, appSeq map (item => if (item.data.installedFromGooglePlay) 1 else 0), IntDataType))
 
@@ -155,8 +155,8 @@ trait EmptyAppMockCursor
     (category, 4, Seq.empty, StringDataType),
     (imagePath, 5, Seq.empty, StringDataType),
     (colorPrimary, 6, Seq.empty, StringDataType),
-    (dateInstalled, 7, Seq.empty, DoubleDataType),
-    (dateUpdate, 8, Seq.empty, DoubleDataType),
+    (dateInstalled, 7, Seq.empty, LongDataType),
+    (dateUpdate, 8, Seq.empty, LongDataType),
     (version, 9, Seq.empty, StringDataType),
     (installedFromGooglePlay, 10, Seq.empty, IntDataType))
